move `QuestionFooter`
