+++ src/app/player/player.module.ts
@@ -13,8 +13,8 @@ import { SharedModule } from '../shared/shared.module';
 import { PlayerRoutingModule } from './player-routing.module';
 import { PlayerComponent } from './player.component';
 import { PlayerService } from './player.service';
+import { QuestionFooterComponent } from './question-footer/question-footer.component';
 import { OneTrueComponent } from './question-screen/one-true/one-true.component';
-import { QuestionFooterComponent } from './question-screen/question-footer/question-footer.component';
 import { QuestionScreenComponent } from './question-screen/question-screen.component';
 import { SeveralTrueComponent } from './question-screen/several-true/several-true.component';
 import { WriteAnswerComponent } from './question-screen/write-answer/write-answer.component';
