+++ src/app/player/question-footer/question-footer.component.html
